opt/apps/server-a/code/sms-backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
pymongo>=4.6,<5.0
PyJWT>=2.8,<3.0
python-dateutil>=2.8,<3.0
gunicorn>=21.2,<22.0
django-cors-headers>=4.4,<5.0
